requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# api/admin.py: 1,2,3
# api/apps.py: 1
# api/migrations/0001_initial.py: 4,5
# api/migrations/0002_auto_20190320_1429.py: 4,5
# api/models/profile.py: 1,2,3,4
# api/serializers.py: 3
# api/tests.py: 1
# api/urls.py: 2
# api/views.py: 5
# config/apps.py: 1
# config/urls.py: 2
# config/wsgi.py: 12
# manage.py: 8
Django == 2.1.7

# api/admin.py: 4
# api/urls.py: 3
api == 0.0.7

# api/permissions.py: 1
# api/serializers.py: 1
# api/urls.py: 1
# api/views.py: 1,4
djangorestframework == 3.9.1

# config/urls.py: 3
djangorestframework_jwt == 1.11.0

allauth